Incorporate feedback from https://github.com/dotnet/corefx/pull/25990 for ProcessStartInfo.StandardInputEncoding feature.
